// src/sdk/generated/mlb.d.ts
export interface paths {
// ...
    "/v3/mlb/scores/{format}/Stadiums": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Stadiums
         * @description Returns all stadiums in the league with capacity, surface, latitude/longitude, city and state (and where applicable country.)
         */
        get: operations["mlb_v3_scores_stadiums"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/v3/mlb/scores/{format}/GamesByDate/{date}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Games - by Date [Live & Final]
         * @description Full scores and gameday info delivered live and post-game. Live data includes innings, scores, pitch count etc. Gameday info includes referee, weather, TV channel etc.
         */
        get: operations["mlb_v3_scores_games___by_date__live___final"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/v3/mlb/odds/{format}/BettingMarketsByGameID/{gameID}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Betting Markets - by Game
         * @description Returns the markets of all available types (e.g. Player Props, Team Props) and available outcomes for a given GameID.
         */
        get: operations["mlb_v3_odds_betting_markets___by_game"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/v3/mlb/odds/{format}/BettingMarketsByGameID/{gameID}/{sportsbookGroup}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Betting Markets - by Game [Sportsbook Group]
         * @description Returns the markets of all available types (e.g. Player Props, Team Props) and available outcomes for a given GameID. A Sportsbook Group must be specified as a URL parameter.
         */
        get: operations["mlb_v3_odds_betting_markets___by_game__sportsbook_group"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/v3/mlb/odds/{format}/PreGameOddsByDate/{date}/{sportsbookGroup}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Pre-Game and Period Game Odds - by Date [Sportsbook Group]
         * @description Returns the full-game and non-full-game odds (spread, moneyline, total) for games on a given date. Only returns the most recently seen odds, not inclusive of line movement. A Sportsbook Group must be specified as a URL parameter.
         */
        get: operations["mlb_v3_odds_pre_game_and_period_game_odds___by_date__sportsbook_group"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/v3/mlb/odds/{format}/PreGameOddsLineMovement/{gameid}/{sportsbookGroup}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Pre-Game and Period Game Odds Line Movement [Sportsbook Group]
         * @description Returns the game odds ( for games in a given week and season. In this endpoint both full-game and partial-game odds are included. Returns the full line movement for the given game. This endpoint has a longer cache as it is meant for historical data/line movement rather than the most up to the second line. A Sportsbook Group must be specified as a URL parameter.
         */
        get: operations["mlb_v3_odds_pre_game_and_period_game_odds_line_movement__sportsbook_group"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/v3/mlb/odds/{format}/InGameOddsByDate/{date}/{sportsbookGroup}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * In-Game Odds - by Date [Sportsbook Group]
         * @description Returns in-play game odds (spread, moneyline, total) for games on a given date. Only returns the most recently seen odds, not inclusive of line movement. As this is in-game, it will only return results while the game is in progress. A Sportsbook Group must be specified as a URL parameter.
         */
        get: operations["mlb_v3_odds_in_game_odds___by_date__sportsbook_group"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/v3/mlb/odds/{format}/InGameOddsLineMovement/{gameid}/{sportsbookGroup}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * In-Game Odds Line Movement [Sportsbook Group]
         * @description Returns in-play game odds (spread, moneyline, total) for games in a given week and season. Returns the full line movement for the given game. This endpoint has a longer cache as it is meant for historical data/line movement rather than the most up to the second line. A Sportsbook Group must be specified as a URL parameter.
         */
        get: operations["mlb_v3_odds_in_game_odds_line_movement__sportsbook_group"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/v3/mlb/stats/{format}/BoxScore/{gameid}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Box Score [Live & Final]
         * @description Full statistical information for a specified game, down to the team and player stat level, delivered live during the game, called per individual game.
         */
        get: operations["mlb_v3_stats_box_score__live___final"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/v3/mlb/pbp/{format}/PlayByPlay/{gameid}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Play By Play [Live & Final]
         * @description Each invididual play, its type and outcome, complete with player and team stats down to the play level, delivered live in real-time, called by game.
         */
        get: operations["mlb_v3_pbp_play_by_play__live___final"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}